# backend/docs/swagger.yaml
definitions:
...
  entity.Key:
    properties:
      id:
        example: 1
        type: integer
      publicKey:
        example: GCK...
        type: string
      walletId:
        example: 1
        type: integer
      weight:
        example: 3
        type: integer
    type: object
...
  entity.Vault:
    properties:
      active:
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      vault_category:
        $ref: '#/definitions/entity.VaultCategory'
      wallet:
        $ref: '#/definitions/entity.Wallet'
    type: object
  entity.VaultCategory:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      theme:
        example: blue
        type: string
    type: object
  entity.Wallet:
    properties:
      funded:
        type: boolean
      id:
        example: 1
        type: integer
      key:
        $ref: '#/definitions/entity.Key'
      type:
        example: sponsor
        type: string
    type: object
...
  v1.CreateVaultCategoryRequest:
    properties:
      name:
        example: Treasury
        type: string
      theme:
        example: blue
        type: string
    required:
    - name
    type: object
  v1.CreateVaultRequest:
    properties:
      assets_id:
        items:
          type: integer
        type: array
      name:
        example: Treasury
        type: string
      vault_category_id:
        example: 1
        type: integer
    required:
    - assets_id
    - name
    - vault_category_id
    type: object
...
  v1.UpdateVaultCategoryRequest:
    properties:
      name:
        example: Vault 2
        type: string
      vault_category_id:
        example: 1
        type: integer
    required:
    - vault_category_id
    type: object
  v1.response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /vault:
    post:
      consumes:
      - application/json
      description: Create and issue a new asset on Stellar
      parameters:
      - description: CreateVaultRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateVaultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new vault
      tags:
      - Vault
...
  /vault-category:
    get:
      consumes:
      - application/json
      description: Get all vault categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.VaultCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all vault categories
      tags:
      - Vault category
    post:
      consumes:
      - application/json
      description: Create and issue a new asset on Stellar
      parameters:
      - description: Vault Category info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateVaultCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.VaultCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new vault category
      tags:
      - Vault Category
  /vault-category/{id}:
    put:
      consumes:
      - application/json
      description: Update a vault category by providing the Vault Category ID and
        the updated information.
      parameters:
      - description: Vault Category ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Vault Category info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateVaultCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated vault category information
          schema:
            $ref: '#/definitions/entity.Vault'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Vault category not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to update vault category'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update a vault category
      tags:
      - Vault Category
...
  /vault/{id}:
    get:
      consumes:
      - application/json
      description: Get vault
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Vault'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get vault
      tags:
      - Vault
...
swagger: "2.0"